+++ packages/geo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lakeql-geo
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - lakeql-core@0.5.0
+
 ## 0.1.6
 
 ### Patch Changes