+++ packages/opfs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lakeql-opfs
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies
+  - lakeql-core@0.5.0
+
 ## 0.0.7
 
 ### Patch Changes